+++ impl/ascend/functions_ext/token_softmax_reducev_inference.cpp
@@ -7,9 +7,9 @@
 #include <cstdlib>
 #include <vector>
 
-#include "../../../adaptor/csrc/impl_functions.hpp"
-#include "../aclnn/acl_scalar.hpp"
 #include "../aclnn/adaptor.hpp"
+#include "../common/acloprunner.hpp"
+#include "impl_functions.hpp"
 
 namespace impl {
 namespace ascend {
@@ -69,12 +69,12 @@ diopiError_t diopiTokenSoftmaxReduceVInference(diopiContextHandle_t ctx, diopiTe
         diopiScalar_t sliceIndexScalar = constructDiopiScalarT(diopi_dtype_int32, i);
         AscendTensor sliceIndexAt;
         makeTensorFromScalar(ctx, sliceIndexAt, &sliceIndexScalar, bLocAt.device());
-        DIOPI_ASCEND_CALL_ACLNN(aclnnIndexSelect, ctx, bLocAt, (int64_t)0, sliceIndexAt, bLocAtSlice);
+        DIOPI_ASCEND_CALL_ACLNN(aclnnIndexSelect, ctx, bLocAt, 0, sliceIndexAt, bLocAtSlice);
         bLocAtSlice.view({bLocAt.shape(1)});
 
         AscendTensor vLocAt;
         makeTensor(ctx, vLocAt, {curSeqLen}, bLocAt.dtype());
-        DIOPI_ASCEND_CALL_ACLNN(aclnnIndexSelect, ctx, bLocAtSlice, (int64_t)0, indexAt, vLocAt);
+        DIOPI_ASCEND_CALL_ACLNN(aclnnIndexSelect, ctx, bLocAtSlice, 0, indexAt, vLocAt);
 
         diopiTensorHandle_t vIndexOut;
         diopiConstTensorHandle_t indexAtHandle = vLocAt.tensorHandle();
